移植初步成功版本
